test: export decodeSourceMappingsLine from expectBundled and reuse it
Addresses review: the VLQ decoder was duplicated between the new
edgecase test and the generic external-sourcemap ordering check.
